Refactor: Move install logic to Mix task and optimize command execution
This refactoring improves code organization and performance:

**1. Moved install logic to proper location**
- Moved all installation code from ExTauri module to Mix.Tasks.ExTauri.Install
- Install task now contains all install-specific logic:
  - cargo_toml/2 - Generates Cargo.toml
  - main_src/3 - Generates Rust main.rs
  - capabilities_json/0 - Generates Tauri V2 capabilities
  - extract_cli_version/1 and build_cli_install_args/1 - CLI helpers
- Module attributes @arg_names and @config_keys moved to Install task

**2. Simplified ExTauri module**
The core module now only contains:
- Application lifecycle (start/2)
- latest_version/0 - Returns latest Tauri version
- installation_path/0 - Returns Tauri CLI installation path
- run/1 - Runs commands that need sidecar (dev, build)
- run_simple/1 - Runs commands without building (info, icon, signer)
- wrap/0 - Builds Burrito release (private)

**3. Optimized command execution**
- Created run_simple/1 for commands that don't need release builds
- info, icon, and signer tasks now use run_simple/1
- This avoids unnecessary Burrito builds for informational commands
- Significant performance improvement for these operations

**Benefits:**
- Better separation of concerns
- Reduced code in core module (580 lines → 146 lines)
- Faster execution for info/icon/signer commands
- More maintainable codebase
- Each task is self-contained with its own logic

**Migration notes:**
- No breaking changes for users
- All Mix tasks work exactly as before
- Internal API changed (ExTauri.install removed, use Mix task directly)
